Major functionality & updates: Checkout cart & process order
Added tests and code, made lots of updates for tests and code for other services too (OrderItem, Coupon, Payment...). Made some entity changes and did some code cleanup.

+++ domain/src/main/java/com/jovana/entity/order/payment/OrderPaymentFailedException.java
@@ -5,6 +5,8 @@ import com.jovana.exception.ExceptionCode;
 import com.jovana.exception.TddBoutiqueApiException;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
+import java.math.BigDecimal;
+
 import static org.springframework.http.HttpStatus.BAD_REQUEST;
 
 /**
@@ -14,16 +16,16 @@ import static org.springframework.http.HttpStatus.BAD_REQUEST;
 @ErrorCode(ExceptionCode.ORDER_PAYMENT_FAILED)
 public class OrderPaymentFailedException extends TddBoutiqueApiException {
 
-    private final Long amount;
+    private final BigDecimal amount;
     private final String description;
 
-    public OrderPaymentFailedException(Long amount, String description, String message) {
+    public OrderPaymentFailedException(BigDecimal amount, String description, String message) {
         super(message);
         this.amount = amount;
         this.description = description;
     }
 
-    public Long getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
